requirements.txt
python-dotenv
requests
flask
jsonify
mysql-connector-python
email-validator
bcrypt
gunicorn
Jinja2
apscheduler